made changes marked with thumbs up + removed token from command line
