+++ cli/docker-compose.yml
@@ -7,12 +7,11 @@ services:
       - "8200:8200"
     restart: always
     volumes:
-      - ./volumes/logs:/vault/logs
-      - ./volumes/file:/vault/file
-      - ./volumes/config:/vault/config
+      - ./volumes/config.hcl:/vault/config.hcl
+      - vault-data:/vault/data
     cap_add:
       - IPC_LOCK
-    entrypoint: vault server -config=config.hcl
+    entrypoint: vault server -config=/vault/config.hcl
 
   keycloak:
     image: quay.io/keycloak/keycloak:16.1.1
@@ -21,4 +20,5 @@ services:
     environment:
       - KEYCLOAK_USER=admin
       - KEYCLOAK_PASSWORD=admin
-
+volumes:
+  vault-data:
