+++ cli/secretFile.txt
@@ -1,0 +1,1 @@
+s.gMyPyhzNq92FnFjWFA6PvK1v
