+++ cli/settings.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/hashicorp/vault/api"
+)
+
+var DEFAULT_CONFIG = api.DefaultConfig()
+var VAULT_ADDRESS = "http://127.0.0.1:8200"
+var TOKEN = getToken()
